chore: refactor UpstreamForm into smaller components
